+++ requirements.txt
@@ -10,9 +10,6 @@ PyPDF2==3.0.1
 tenacity==8.2.3
 langchain==0.2.7
 langchain-openai==0.1.16
-langchain-groq==0.1.3
-langchain-anthropic==0.1.11
-langchain-google-genai==1.0.3
 langchain-huggingface==0.0.3
 faiss-cpu==1.8.0
 tika==2.6.0
