+++ scraper/config.py
@@ -41,19 +41,6 @@ tasks = [
     ),
 ]
 
-# Matches the embedding model names with the language
-embedding_models_dict = {
-    "OpenAI": "text-embedding-ada-002",
-}
-
-
-class LLMConfig(BaseModel):
-    model_name: str
-    api_key: str = None
-    embedding_model_name: str = ""
-    temperature: Optional[float] = Field(0.7, ge=0.0, le=1.0)
-    max_tokens: Optional[int] = Field(1_000, gt=0)
-
 
 class AnswerItem(BaseModel):
     text: str = Field(
